Abstracted logic out of StoneMacro into StoneOutputBlock
