+++ src/Stone/Types/index.js
@@ -1,8 +1,9 @@
 export const Types = {
 	StoneDump: require('./StoneDump'),
 	StoneEmptyExpression: require('./StoneEmptyExpression'),
-	StoneOutput: require('./StoneOutput'),
-	StoneOutputExpression: require('./StoneOutputExpression'),
 	StoneLoop: require('./StoneLoop'),
 	StoneMacro: require('./StoneMacro'),
+	StoneOutput: require('./StoneOutput'),
+	StoneOutputBlock: require('./StoneOutputBlock'),
+	StoneOutputExpression: require('./StoneOutputExpression'),
 }
